+++ packages/charts/src/utils/themes/theme.ts
@@ -324,6 +324,19 @@ export interface MetricStyle {
   minHeight: Pixels;
 }
 
+/** @alpha */
+export interface FlamegraphStyle {
+  navigation: {
+    textColor: Color;
+    buttonTextColor: Color;
+    buttonDisabledTextColor: Color;
+    buttonBackgroundColor: Color;
+    buttonDisabledBackgroundColor: Color;
+  };
+  scrollbarTrack: Color;
+  scrollbarThumb: Color;
+}
+
 /** @public */
 export interface ScalesConfig {
   /**
@@ -484,6 +497,9 @@ export interface Theme {
    * Theme styles for tooltip
    */
   tooltip: TooltipStyle;
+
+  /** @alpha */
+  flamegraph: FlamegraphStyle;
 }
 
 /** @public */