+++ storybook/stories/icicle/04_cpu_profile_gl_flame.story.tsx
@@ -125,7 +125,3 @@ export const Example = () => {
     </Chart>
   );
 };
-
-Example.parameters = {
-  background: { default: 'white' },
-};